% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching_utils.R
\name{add_random_na}
\alias{add_random_na}
\title{add_random_na}
\usage{
add_random_na(data, n, exclude_cols = NULL)
}
\arguments{
\item{data}{a dataframe}

\item{n}{number of NA values to insert}

\item{exclude_cols}{string or vector of strings with column to exclude
from NA filling}
}
\value{
a dataframe with random NA
}
\description{
put random NA values within a dataframe for teaching and
testing purposes.
}
