% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{all_same}
\alias{all_same}
\title{all_same}
\usage{
all_same(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{logical indicating whether \code{NA} values should be removed or not.}
}
\value{
logical vector
}
\description{
Check if all values in a vector are the same. It is robust also to NA values
that can be handled with the \code{na.rm} parameter
}
\examples{
all_same(c(1,1,1))  # TRUE
all_same(c(1,2,3))  # FALSE
all_same(c(1,1,NA)) # FALSE
all_same(c(1,1,NA), na.rm = TRUE) # TRUE
}
