% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_utils.R
\name{cap_link}
\alias{cap_link}
\title{cap_link}
\usage{
cap_link(link, text = NULL)
}
\arguments{
\item{link}{character. The web address}

\item{text}{charactet. The text to be displayed. Default to \code{NULL} thus
the same as the \code{link}}
}
\value{
The html code for the hyperlink
}
\description{
Format html code for an hyperlink
}
