% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_utils.R
\name{cinorm}
\alias{cinorm}
\title{cinorm}
\usage{
cinorm(x = NULL, se = NULL, avg = NULL, interval = 0.95)
}
\arguments{
\item{x}{a vector}

\item{se}{the standard error}

\item{avg}{the mean}

\item{interval}{the desired interval}
}
\value{
the upper and lower CI bound
}
\description{
cinorm
}
