% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_utils.R
\name{closest_number}
\alias{closest_number}
\title{closest_number}
\usage{
closest_number(vector, target)
}
\arguments{
\item{vector}{numeric}

\item{target}{numeric}
}
\value{
numeric
}
\description{
Find the closest number in a vector given
a target number
}
