% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_utils.R
\name{col_has_na}
\alias{col_has_na}
\title{col_has_na}
\usage{
col_has_na(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
character vector
}
\description{
Return columns names where there is
at least one NA value
}
