% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{collapse_rows}
\alias{collapse_rows}
\title{Collapse rows of a dataframe}
\usage{
collapse_rows(data, cols, to = NA)
}
\arguments{
\item{data}{a dataframe}

\item{cols}{character vector with columns to collapse}

\item{to}{value to put. Default to `NA`}
}
\value{
a dataframe
}
\description{
Collapse rows of a dataframe
}
\examples{
dt <- data.frame(a = c(1, 1, 2, 2), b = c("a", "a", "a", "b"))
collapse_rows(dt, "a", to = "")
}
