% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_utils.R
\name{colorize}
\alias{colorize}
\title{colorize}
\usage{
colorize(x, color)
}
\arguments{
\item{x}{the text to be formatted}

\item{color}{the color to be applied}
}
\value{
string with the formatted p value
}
\description{
thanks to https://bookdown.org/yihui/rmarkdown-cookbook/font-color.html#using-an-r-function-to-write-raw-html-or-latex-code
}
\examples{
colorize("red text", "red")
}
