% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{compare_rma}
\alias{compare_rma}
\title{compare_rma}
\usage{
compare_rma(..., fitlist = NULL, extra_params = NULL)
}
\arguments{
\item{...}{models to be compared}

\item{fitlist}{a list of models instead of using `...`. When `fitlist` is provided `...` are ignored.}

\item{extra_params}{a vector of (extra) parameters to be extracted. These are the elements of the \code{rma.uni} objects.}
}
\value{
a dataframe
}
\description{
Put in a dataframe one or more models fitted with \code{metafor::rma.uni} for a nice visual comparison.
}
