% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{conditional}
\alias{conditional}
\title{conditional}
\usage{
conditional(fun)
}
\arguments{
\item{fun}{function}
}
\description{
Function to create a wrapper version of the input function
that run according to a logical condition. Useful for conditional pipelines.
see https://community.rstudio.com/t/conditional-pipelines/6076/2
}
