% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_utils.R
\name{coridx}
\alias{coridx}
\title{coridx}
\usage{
coridx(p)
}
\arguments{
\item{p}{integer for the size of the correlation matrix}
}
\value{
a character vector with all the correlations
}
\description{
create the correlation names given the size of the correlation matrix
}
