% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{create_R_env}
\alias{create_R_env}
\title{create_R_env}
\usage{
create_R_env(open = TRUE)
}
\arguments{
\item{open}{logical indicating if opening the file interactively}
}
\description{
Create an ".Renviron" file if not exists and optionally opens it.
}
