% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_fun.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{deg_to_rad}
\usage{
deg_to_rad(deg)
}
\arguments{
\item{deg}{numeric indicating the angle in degrees}
}
\value{
the angle in degrees
}
\description{
convert degrees to radians
}
