% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_utils.R
\name{disp_file}
\alias{disp_file}
\title{disp_file}
\usage{
disp_file(file)
}
\arguments{
\item{file}{string that indicate the path and filename}
}
\value{
Invisibly return the file lines and print the content
}
\description{
disp_file
}
