% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_fun.R
\name{dur_to_frames}
\alias{dur_to_frames}
\title{dur_to_frames}
\usage{
dur_to_frames(duration, refresh_rate)
}
\arguments{
\item{duration}{the duration in ms}

\item{refresh_rate}{the monitor refresh rate (e.g. 60)}
}
\value{
the number of frames
}
\description{
return the number of frames associated with a specific duration (in ms)
}
