% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{edit_R_env}
\alias{edit_R_env}
\title{edit_R_env}
\usage{
edit_R_env(x)
}
\arguments{
\item{x}{a named vector}
}
\description{
Edit an .Renviron file providing a named vector in the form key=value.
Existing keys will be overwritten and new keys will be added.
}
