% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_utils.R
\name{epredict}
\alias{epredict}
\title{epredict}
\usage{
epredict(model, newdata, ...)
}
\arguments{
\item{model}{a fitted glm/lm object}

\item{newdata}{a list or dataframe with values where evaluate the prediction}

\item{...}{additional arguments to \code{predict()}}
}
\description{
improved predict function, it works as the standard predict function but prints the model equation that is supporting the current prediction
}
