% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching_utils.R
\name{extract_pdf_pages}
\alias{extract_pdf_pages}
\title{extract_pdf_pages}
\usage{
extract_pdf_pages(file, pages, out = NULL)
}
\arguments{
\item{file}{character with the pdf file}

\item{pages}{integer vector with pdf pages to extract}

\item{out}{character with the output file name. Default to NULL}
}
\description{
extract_pdf_pages
}
