% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_utils.R
\name{fac2con}
\alias{fac2con}
\title{fac2con}
\usage{
fac2con(x)
}
\arguments{
\item{x}{a factor with two levels}
}
\description{
Convert a factor to the contrasts representation. This works only with factors with two levels.
}
