% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_utils.R
\name{filter_output}
\alias{filter_output}
\title{filter_output}
\usage{
filter_output(x, lines = NULL, cat = TRUE)
}
\arguments{
\item{x}{the result of a function to capture e.g. summary(x)}

\item{lines}{a vector of line numbers to extract or a pattern as "from|to" where from and to are regex mapping the lines to extract.}

\item{cat}{logical indicating if using cat() for the output}
}
\description{
filter_output
}
\examples{
fit <- lm(mpg ~ hp, data = mtcars)
filter_output(summary(fit), 1:5)
}
