% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{find_pattern}
\alias{find_pattern}
\title{find_pattern}
\usage{
find_pattern(pattern, which_files = c("R", "Rmd"), dir = ".")
}
\arguments{
\item{which_files}{a character vector with file extensions to check.}

\item{dir}{starting directory from which an iterative search
is performed}

\item{fun}{string or name with the pattern to search}
}
\description{
return how many times a specific pattern occurs in any file of
the current directory and sub directories.
}
