% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_utils.R
\name{fpvalue}
\alias{fpvalue}
\title{fpvalue}
\usage{
fpvalue(p, limit = 0.001, tex = FALSE)
}
\arguments{
\item{p}{numeric. The p value to format}

\item{limit}{numeric. the minimum value to print. Default to 0.001}

\item{tex}{logical. If TRUE, the result is inserted into two \code{$...$}}
}
\value{
string with the formatted p value
}
\description{
fpvalue
}
\examples{
fpvalue(0.001, tex = TRUE)
fpvalue(0.001, tex = FALSE)
}
