% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_fun.R
\name{frames_to_dur}
\alias{frames_to_dur}
\title{frames_to_dur}
\usage{
frames_to_dur(nframes, refresh_rate)
}
\arguments{
\item{nframes}{the number of frames}

\item{refresh_rate}{the monitor refresh rate (e.g. 60)}
}
\value{
the duration in ms
}
\description{
frames_to_dur
}
