% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_utils.R
\name{get_bib_entries}
\alias{get_bib_entries}
\title{get_bib_entries}
\usage{
get_bib_entries(file = knitr::current_input())
}
\arguments{
\item{file}{character that indicate the input file. Default to
the current \code{.Rmd} file if not specified.}
}
\value{
character vector with all bibtex entry used in a document
}
\description{
get_bib_entries
}
