% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_utils.R
\name{get_caption}
\alias{get_caption}
\title{get_caption}
\usage{
get_caption()
}
\value{
The pandoc syntax for the cross-referencing
}
\description{
Match a reference within an R Markdown document using the chunk label
}
