% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_utils.R
\name{get_chunk_labels}
\alias{get_chunk_labels}
\title{get_chunk_labels}
\usage{
get_chunk_labels(file, output = NULL, as_captions = FALSE)
}
\arguments{
\item{file}{the input rmd document}

\item{output}{the output text document. Default to NULL}

\item{as_captions}{logical. If the output should be formatted as captions}
}
\description{
get_chunk_labels
}
