% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_utils.R
\name{get_funs}
\alias{get_funs}
\title{get_funs}
\usage{
get_funs(file)
}
\arguments{
\item{file}{an .R file from which extracting function}
}
\value{
a named list with function
}
\description{
Extract functions from an .R file as text, creating a named list.
}
