% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_utils.R
\name{get_model_formula}
\alias{get_model_formula}
\title{get_model_formula}
\usage{
get_model_formula(fit)
}
\arguments{
\item{fit}{a model object}
}
\value{
A character with the model formula
}
\description{
get_model_formula
}
