% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_utils.R
\name{get_na_rows}
\alias{get_na_rows}
\title{get_na_rows}
\usage{
get_na_rows(data, ..., subset = FALSE)
}
\arguments{
\item{data}{dataframe}

\item{...}{comma separated names of columns to evaluate}

\item{subset}{logical that indicate if returning the subsetted data.
Default to \code{false}}
}
\value{
integer vector
}
\description{
Return the rows index where there are NA values in specific
columns
}
