% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_fun.R
\name{get_stim_degree}
\alias{get_stim_degree}
\title{get_stim_degree}
\usage{
get_stim_degree(size, distance)
}
\arguments{
\item{size}{stimulus size}

\item{distance}{distance from the screen}
}
\value{
the angle
}
\description{
calculating the visual
}
