% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_fun.R
\name{get_stim_size}
\alias{get_stim_size}
\title{get_stim_size}
\usage{
get_stim_size(vis_angle, distance)
}
\arguments{
\item{vis_angle}{stimulus size in degrees of visual angles}

\item{distance}{distance from the screen}
}
\value{
the distance
}
\description{
calculating the visual
}
