% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{gh_down_link}
\alias{gh_down_link}
\title{gh_down_link}
\usage{
gh_down_link(x, hyperlink = FALSE, url_name = NULL)
}
\arguments{
\item{x}{character with the github link}

\item{url_name}{character with the url name for the hyperlink}

\item{print_url}{logical}
}
\description{
create a download/view link for a github file from url
}
