% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{I2}
\alias{I2}
\title{I2}
\usage{
I2(tau, vi)
}
\arguments{
\item{tau}{the standard error (tau)}

\item{vi}{the typical sampling variability}
}
\value{
numeric the I^2 value
}
\description{
I2
}
\examples{
I2(0.1, 0.08)
}
