% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{is_valid_url}
\alias{is_valid_url}
\title{is_valid_url}
\usage{
is_valid_url(string)
}
\arguments{
\item{string}{}
}
\value{
logical
}
\description{
Check if a string is a valid url. Thanks to https://stackoverflow.com/a/73952264/9032257
}
