% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_utils.R
\name{leave1out}
\alias{leave1out}
\title{leave1out}
\usage{
leave1out(data, group_var = NULL, keep_full = FALSE)
}
\arguments{
\item{data}{a dataframe}

\item{group_var}{name of the variable that identify each
observation to remove. If \code{NULL} will be removed one
dataframe line at time.}
}
\value{
a named list of dataframes
}
\description{
leave1out
}
