% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{letter}
\alias{letter}
\title{Output formats for filor}
\usage{
letter(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{pdf_document}.}
}
\value{
An R Markdown output format object.
}
\description{
Output formats for filor
}
