% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_utils.R
\name{lme4_has_converged}
\alias{lme4_has_converged}
\title{Check lme4 Model Convergence Status}
\usage{
lme4_has_converged(x)
}
\arguments{
\item{x}{An object of class `merMod` (e.g., the result of `lmer()` or `glmer()`).}
}
\value{
An integer indicating the convergence status:
\describe{
  \item{1}{Converged successfully (no convergence issues).}
  \item{0}{Model is singular (some random effects may be redundant).}
  \item{-1}{Convergence issues detected, but model is not singular.}
}
}
\description{
`lme4_has_converged` checks whether a model fitted with `lme4::lmer` or `lme4::glmer` has converged successfully.
}
\details{
This function inspects the `@optinfo$conv$lme4` slot of the model object to determine if
the optimizer reported any convergence problems. It also checks for singularity using
`isSingular()`. The function has been copied from Robert Long https://stackoverflow.com/a/72128391/9032257.
}
\examples{
\dontrun{
library(lme4)
fm <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
lme4_has_converged(fm)
}

}
