% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_utils.R
\name{model_equation}
\alias{model_equation}
\title{model_equation}
\usage{
model_equation(model, newdata = NULL, ..., only_print = FALSE)
}
\arguments{
\item{model}{a fitted glm/lm object}

\item{newdata}{a list or dataframe with values where evaluate the prediction}

\item{...}{additional arguments to \code{predict()}}

\item{only_print}{logical. TRUE if returning the equation as string}
}
\description{
print model equation with generic or actual values thanks to https://stats.stackexchange.com/a/433060 for the equation part. I've added the printing and prediction
}
