% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_utils.R
\name{mseq}
\alias{mseq}
\title{mseq}
\usage{
mseq(from = 1, to = 1)
}
\arguments{
\item{from}{vector of integers for starting points of the sequence. Default to 1}

\item{to}{vector of integers for ending points of the sequence. Default to 1}
}
\value{
a vector of integers with all the sequences
}
\description{
create a sequence with multiple from and to vectors
}
