% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_utils.R
\name{named_list}
\alias{named_list}
\title{named_list}
\usage{
named_list(...)
}
\arguments{
\item{...}{comma separated R objects}
}
\value{
a named list
}
\description{
Create a named list with input objects directly using the object name of the global environment
}
