% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_utils.R
\name{ncor}
\alias{ncor}
\title{ncor}
\usage{
ncor(p)
}
\arguments{
\item{p}{integer for the size of the correlation matrix}
}
\value{
the number of correlations
}
\description{
calculate the number of correlations given the size of the correlation matrix. In practice is the same as counting the number of elements in the upper/lower triangle of the correlation matrix
}
