% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_utils.R
\name{odds_ratio}
\alias{odds_ratio}
\title{odds_ratio}
\usage{
odds_ratio(pn, pd)
}
\arguments{
\item{pn}{numeric the probability at the numerator}

\item{pd}{numeric the probability at the denominator}
}
\value{
the odds ratio
}
\description{
odds_ratio
}
