% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{pb}
\alias{pb}
\title{pb}
\usage{
pb(index, niter, width = 40)
}
\arguments{
\item{index}{Integer that indicate the current iteration}

\item{niter}{Integer that indicate the total number of iterations}

\item{width}{Maximum width of the console output}
}
\description{
Display a progress bar to use within a \code{for} loop
}
