% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching_utils.R
\name{purl_here}
\alias{purl_here}
\title{purl_here}
\usage{
purl_here(file, output = NULL)
}
\arguments{
\item{file}{character with the \code{.Rmd} file to purl}

\item{output}{character with the path and filename of the output file. Default to the \code{.Rdm} file path.}
}
\description{
purl_here
}
