% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{r_to_z}
\alias{r_to_z}
\title{r_to_z}
\usage{
r_to_z(r)
}
\arguments{
\item{r}{the correlation}
}
\value{
the Fisher's z transformed correlation
}
\description{
convert a correlation into the Fisher's z
}
