% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_fun.R
\name{rad_to_deg}
\alias{rad_to_deg}
\title{rad_to_deg}
\usage{
rad_to_deg(rad)
}
\arguments{
\item{rad}{numeric indicating the angle in radians}
}
\value{
the angle in radians
}
\description{
convert radians to degrees
}
