% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_utils.R
\name{rdim}
\alias{rdim}
\title{rdim}
\usage{
rdim(k)
}
\arguments{
\item{k}{the number of unique correlations}
}
\value{
the size of the correlation matrix
}
\description{
Calculate the size of the correlation matrix given the number of unique correlations
}
