% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_utils.R
\name{remove_axes}
\alias{remove_axes}
\title{remove_axis}
\usage{
remove_axes(where, what)
}
\arguments{
\item{where}{character vector indicating which axis to remove (e.g., "x" or c("x", "y"))}

\item{what}{character vector indicating which elements to remove (e.g., "title")}
}
\description{
Utility to remove in a vectorized manner axis elements from a \code{ggplot2} object
}
