% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_utils.R
\name{remove_revision}
\alias{remove_revision}
\title{remove_revision}
\usage{
remove_revision(file, out = NULL)
}
\arguments{
\item{file}{the input file}

\item{out}{the output file name. Default to \code{NULL}}
}
\description{
remove_revision
}
