% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_utils.R
\name{rename_figs}
\alias{rename_figs}
\title{rename_figs}
\usage{
rename_figs(prefix = "figure")
}
\arguments{
\item{prefix}{character to be placed before the actual name. Default to "figure"}
}
\description{
Knitr hook for creating a copy of figures named in progressive order (e.g., Figure 1) instead of the chunk label
}
