% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching_utils.R
\name{sampdist}
\alias{sampdist}
\title{sampdist}
\usage{
sampdist(n, mean = 0, sd = 1, B = 1000, FUN)
}
\arguments{
\item{n}{sample size}

\item{mean}{the mean of the normal distribution}

\item{sd}{the standard deviation of the normal distribution}

\item{B}{the number of simulations}

\item{FUN}{the function to be applied to each sample}
}
\value{
a vector of length B with the sampling distribution
}
\description{
Generate a sampling distribution from a statistics applied to samples generated from a normal distribution.
}
\examples{
sampdist(30, 0, 1, B = 10, FUN = mean)
}
