% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{SE4power}
\alias{SE4power}
\title{Calculate Standard Error (SE) associated with a desired level of statistical power}
\usage{
SE4power(b = NA, power = 0.8, alpha = 0.05, alternative = "two.sided")
}
\arguments{
\item{power}{The desired level of statistical power}

\item{alpha}{The alpha level (rate of type I errors)}

\item{alternative}{Whether the statistical test is two sided or one sided. Must be one of "two.sided" (default) or "one.sided"}

\item{B}{The model parameter B (practically, the effect size)}
}
\value{
The numerical value of the Standard Error (SE) associated with the above arguments
}
\description{
Calculate the Standard Error (SE) that is associated with a given power and a given alpha level for a given value of a model parameter B
}
