% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_utils.R
\name{show_file}
\alias{show_file}
\title{show_file}
\usage{
show_file(file, how = c("bare", "generic", "code"), engine = NULL)
}
\arguments{
\item{file}{string that indicate the path and filename}

\item{how}{if bare, the file is printed as it is. With generic the file is embedded into backticks.
If code, the file is embedded with backticks and an engine. If not specified the default engine is R.}
}
\value{
Invisibly return the file lines and print the content
}
\description{
show_file
}
