% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{sim_meta}
\alias{sim_meta}
\title{sim_meta}
\usage{
sim_meta(k, mu, tau2, n1, n2 = NULL, v = 1)
}
\arguments{
\item{k}{number of studies.}

\item{mu}{average effect size.}

\item{tau2}{between studies heterogeneity.}

\item{n1}{sample size for the first group.}

\item{n2}{sample size for the second group. Default to \code{NULL}}

\item{v}{pooled within-study variance. Default to \code{1}}
}
\value{
dataframe with simulated data
}
\description{
Simulate meta-analysis data using unstandardized mean difference as effect size.
}
\examples{
# simulate equal-effects model with 10 studies, 30 participants in each study
# and an effect size of 0.5
sim_meta(10, 0.5, 0, 30, 30)
}
