% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{sourceR}
\alias{sourceR}
\title{sourceR}
\usage{
sourceR(files = NULL, path = "R")
}
\arguments{
\item{files}{files to source. Default to NULL (i.e., all files)}

\item{path}{path with .R files. Default to "R/" as in standard package-like projects.}
}
\description{
Load one or more R functions from R/ folder (or any custom folder)
}
