% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{summary_rma}
\alias{summary_rma}
\title{summary_rma}
\usage{
summary_rma(x, extra_params = NULL)
}
\arguments{
\item{x}{a model fitted with \code{metafor::rma.uni()}}

\item{extra_params}{a vector of (extra) parameters to be extracted (see details). These are the elements of the \code{rma.uni} objects.}
}
\value{
a dataframe
}
\description{
summary_rma
}
\details{
By default the function extract \code{c("b", "se", "zval", "pval", "ci.lb", "ci.ub")} as parameters.
}
