% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{tau2_from_I2}
\alias{tau2_from_I2}
\title{tau2_from_I2}
\usage{
tau2_from_I2(I2, vi)
}
\arguments{
\item{I2}{the I squared value}

\item{vi}{the effect size sampling variability}
}
\value{
the tau2 value
}
\description{
tau2_from_I2
}
\examples{
tau2_from_I2(0.2, 0.08)
}
