% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{trim}
\alias{trim}
\title{trim}
\usage{
trim(x, min = NULL, max = NULL)
}
\arguments{
\item{x}{the numeric vector}

\item{min}{the minimum value. Default to `NULL`. If `NULL`, the lower bound is not modified.}

\item{max}{the maximum value. Default to `NULL`. If `NULL`, the upper bound is not modified.}
}
\description{
Cut a numeric vector between min and max. Values lower than min or higher than max are substituted with the provided values.
}
\examples{
x <- runif(20, 0, 20)
trim(x, min = 5, max = 10)
trim(x, max = 10)
trim(x, min = 5)
}
