% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching_utils.R
\name{trim_df}
\alias{trim_df}
\title{trim_df}
\usage{
trim_df(data, n = 4, digits = 3)
}
\arguments{
\item{data}{a dataframe}

\item{n}{number of rows to display before and after the dots}
}
\value{
a dataframe
}
\description{
trim_df
}
