% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{try_seed}
\alias{try_seed}
\title{try_seed}
\usage{
try_seed(expr, maxrun = 100, digits = 4)
}
\arguments{
\item{expr}{an R expression}

\item{maxrun}{the maximum number of iterations}

\item{digits}{the number of digits for the random seed}
}
\description{
Repeat an R expression a certain number of times with a random seed. Each iteration is self-paced by pressing a key. The random seed is printed along with the result of the
}
