% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_utils.R
\name{wald_test}
\alias{wald_test}
\title{wald_test}
\usage{
wald_test(x, h0 = 0, btt = NULL, alternative = "two.sided", alpha = 0.05)
}
\arguments{
\item{x}{a model. Currently supported classes are `lmerMod`, `glmerMod`, `lm` and `glm`}

\item{h0}{a vector of null hypothesis, default to 0}

\item{btt}{vector of coefficents names to test. Default to `NULL` (all coefficients)}

\item{alternative}{character indicating the side of the hypothesis}

\item{alpha}{the alpha level, default to 0.05}
}
\value{
a dataframe
}
\description{
wald_test
}
\examples{
fit <- lm(Sepal.Lenght ~ Species, data = iris)
wald_test(fit)
}
