% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{wcor}
\alias{wcor}
\title{Pairwise name combinations (as concatenated strings)}
\usage{
wcor(n, names = NULL)
}
\arguments{
\item{n}{Integer. Number of elements (and number of names to generate if
`names` is `NULL`).}

\item{names}{Character vector of length `n`. If `NULL`, defaults to
`letters[1:n]`.}
}
\value{
A character vector of length `choose(n, 2)` containing concatenated
  name pairs.
}
\description{
Generate all 2-combinations of a set of names and return each pair as a
single concatenated string (e.g., `"ab"`, `"ac"`, ...).
}
\details{
If `names` is not provided, the first `n` lowercase letters (`letters[1:n]`)
are used as default names.


Uses [gtools::combinations()] to generate all unordered pairs of `names`,
then concatenates each pair with no separator.
}
\examples{
wcor(4)
wcor(3, c("x", "y", "z"))

}
\seealso{
[gtools::combinations()]
}
