% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_utils.R
\name{write_bib_rmd}
\alias{write_bib_rmd}
\title{write_bib_rmd}
\usage{
write_bib_rmd(file = knitr::current_input(), input_bib, output_bib)
}
\arguments{
\item{file}{character that indicate the input file}

\item{input_bib}{character that indicate the input \code{.bib} file}

\item{output_bib}{character that indicate the output \code{.bib} file}
}
\value{
the bibfile
}
\description{
write_bib_rmd
}
