% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{compare_rma}
\alias{compare_rma}
\title{compare_rma}
\usage{
compare_rma(...)
}
\arguments{
\item{...}{models of class \code{rma.uni} or \code{rma.mv}}
}
\value{
a dataframe with coefficients
}
\description{
compare_rma
}
