% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_utils.R
\name{get_funs}
\alias{get_funs}
\title{get_funs}
\usage{
get_funs(files)
}
\arguments{
\item{file}{one or multiple .R files as character vector from which extracting function}
}
\value{
a named list with function
}
\description{
Extract functions from .R files as text, creating a named list.
}
