% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{pb}
\alias{pb}
\title{pb}
\usage{
pb(niter, index)
}
\arguments{
\item{niter}{Integer that indicate the total number of iterations}

\item{index}{Integer that indicate the current iteration}
}
\description{
Display a progress bar to use within a \code{for} loop
}
